backend/requirements.txt
fastapi
uvicorn
jina
qdrant-client
redis
python-dotenv
requests
google-generativeai
